+++ spec/support/json_response.rb
@@ -1,0 +1,5 @@
+module JsonResponse
+  def json
+    JSON.parse(response.body)
+  end
+end